revert redundant diff

+++ paddle/phi/kernels/gpu/repeat_interleave_grad_kernel.cu
@@ -227,7 +227,7 @@ PD_REGISTER_KERNEL(repeat_interleave_with_tensor_index_grad,
                    double,
                    int,
                    int64_t,
-                   phi::float16) {}
+                   phi::bfloat16) {}
 PD_REGISTER_KERNEL(repeat_interleave_grad,
                    GPU,
                    ALL_LAYOUT,
@@ -236,4 +236,4 @@ PD_REGISTER_KERNEL(repeat_interleave_grad,
                    double,
                    int,
                    int64_t,
-                   phi::float16) {}
+                   phi::bfloat16) {}
